+++ src/parsing/load_file.c
@@ -25,7 +25,7 @@ static void	saved_update(char **buffer, char **saved)
 int	load_file(char *file_name, char ***tokens)
 {
 	int		fd;
-	int		read_call;
+	ssize_t	read_call;
 	char	*buffer;
 	char	*saved;
